refactor(client/slices): move all slices to a new folder
