+++ client/src/redux/store/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
-import authSliceReducer from "../../authentication/authSlice";
+import authSliceReducer from "../../slices/authentication/authSlice";
 import rootReducer from '../reducer/reducer'
-import { apiSlice } from "../../authentication/apiSlice";
+import { apiSlice } from "../../slices/authentication/apiSlice";
 
 export const store = configureStore({
     reducer:{
